+++ nifi-nar-bundles/nifi-aws-bundle/nifi-aws-schema-registry-service/src/main/java/org/apache/nifi/aws/schemaregistry/client/SchemaRegistryClient.java
@@ -26,5 +26,5 @@ public interface SchemaRegistryClient {
 
     RecordSchema getSchema(final String schemaName) throws IOException, SchemaNotFoundException;
 
-    RecordSchema getSchema(final String schemaName, final int version) throws IOException, SchemaNotFoundException;
+    RecordSchema getSchema(final String schemaName, final long version) throws IOException, SchemaNotFoundException;
 }
